+++ src/Controller/UserController.php
@@ -32,9 +32,16 @@ class UserController extends AbstractController
      * @route("/AfficherUsers", name="AfficherUsers")
      */
     public function AfficherUsers(UserRepository $repository){
-        $users=$repository->findAll();
-        return $this->render('user/afficher.html.twig',
-            ['users'=>$users, 'user' => $this->getUser()->getUsername() ]);
+
+        if($this->getUser()->isAdmin())
+        {
+            $users=$repository->findAll();
+            return $this->render('user/afficher.html.twig',
+                ['users'=>$users, 'user' => $this->getUser()->getUsername() ]);
+        }
+
+        return $this->redirectToRoute('app_login');
+
     }
 
     /**
@@ -43,20 +50,34 @@ class UserController extends AbstractController
      * @route("/ProfilBack/{id}", name="ProfilBack")
      */
     public function ProfilBack($id, UserRepository $repository){
-        $profil=$repository->find($id);
-        return $this->render('user/profilBack.html.twig',
-            ['profil'=>$profil, 'user' => $this->getUser()->getUsername() ]);
+
+        if($this->getUser()->isAdmin())
+        {
+            $profil=$repository->find($id);
+            return $this->render('user/profilBack.html.twig',
+                ['profil'=>$profil, 'user' => $this->getUser()->getUsername() ]);
+        }
+
+        return $this->redirectToRoute('app_login');
+
     }
 
     /**
      * @route("/SupprimerUser/{id}",name="SupprimerUser")
      */
     function SupprimerUser($id,UserRepository $repository){
-        $user=$repository->find($id);
-        $em=$this->getDoctrine()->getManager();
-        $em->remove($user);
-        $em->flush();
-        return $this->redirectToRoute('AfficherUsers');
+
+        if($this->getUser()->isAdmin())
+        {
+            $user=$repository->find($id);
+            $em=$this->getDoctrine()->getManager();
+            $em->remove($user);
+            $em->flush();
+            return $this->redirectToRoute('AfficherUsers');
+        }
+
+        return $this->redirectToRoute('app_login');
+
     }
 
     /**
@@ -66,21 +87,28 @@ class UserController extends AbstractController
      */
     function AjouterUser(Request $request, UserPasswordEncoderInterface $encoder){
 
-        $user = new User();
-        $form=$this->createForm(UserType::class,$user);
-        $form->add('Ajouter',SubmitType::class);
-        $form->handleRequest($request);
-        if($form->isSubmitted() && $form->isValid()){
-            $hash = $encoder->encodePassword($user, $user->getPassword());
-            $user->setPassword($hash);
-            $em=$this->getDoctrine()->getManager();
-            $em->persist($user);
-            $em->flush();
-            return  $this->redirectToRoute('AfficherUsers');
+        if($this->getUser()->isAdmin())
+        {
+            $user = new User();
+            $form=$this->createForm(UserType::class,$user);
+            $form->add('Ajouter',SubmitType::class);
+            $form->handleRequest($request);
+            if($form->isSubmitted() && $form->isValid()){
+                $hash = $encoder->encodePassword($user, $user->getPassword());
+                $user->setPassword($hash);
+                $em=$this->getDoctrine()->getManager();
+                $em->persist($user);
+                $em->flush();
+                return  $this->redirectToRoute('AfficherUsers');
+            }
+            return $this->render('user/ajouter.html.twig',[
+                'form'=>$form->createView(), 'user' => $this->getUser()->getUsername()
+            ]);
         }
-        return $this->render('user/ajouter.html.twig',[
-            'form'=>$form->createView(), 'user' => $this->getUser()->getUsername()
-        ]);
+
+        return $this->redirectToRoute('app_login');
+
+
     }
 
     /**
@@ -88,20 +116,27 @@ class UserController extends AbstractController
      */
     function ModifierUser(UserRepository  $repository, $id, Request $request, UserPasswordEncoderInterface $encoder){
 
-        $user=$repository->find($id);
-        $form=$this->createForm(UserType::class, $user);
-        $form->add('Modifier',SubmitType::class);
-        $form->handleRequest($request);
-        if ($form->isSubmitted() && $form->isValid()){
-            $hash = $encoder->encodePassword($user, $user->getPassword());
-            $user->setPassword($hash);
-            $em=$this->getDoctrine()->getManager();
-            $em->flush();
-            return $this->redirectToRoute("AfficherUsers");
+        if($this->getUser()->isAdmin())
+        {
+            $user=$repository->find($id);
+            $form=$this->createForm(UserType::class, $user);
+            $form->add('Modifier',SubmitType::class);
+            $form->handleRequest($request);
+            if ($form->isSubmitted() && $form->isValid()){
+                $hash = $encoder->encodePassword($user, $user->getPassword());
+                $user->setPassword($hash);
+                $em=$this->getDoctrine()->getManager();
+                $em->flush();
+                return $this->redirectToRoute("AfficherUsers");
+            }
+            return $this->render('user/modifier.html.twig',[
+                'form'=>$form->createView(), 'user' => $this->getUser()->getUsername()
+            ]);
         }
-        return $this->render('user/modifier.html.twig',[
-            'form'=>$form->createView(), 'user' => $this->getUser()->getUsername()
-        ]);
+
+        return $this->redirectToRoute('app_login');
+
+
     }
 
 
@@ -167,7 +202,7 @@ class UserController extends AbstractController
             $user->setPassword($hash);
             $em=$this->getDoctrine()->getManager();
             $em->flush();
-            return $this->redirectToRoute("AfficherProduitsFront");
+            return $this->redirectToRoute("AfficherUsers");
         }
         return $this->render('user/modifierProfil.html.twig',[
             'form'=>$form->createView(), 'user' => $this->getUser()->getUsername()
